Move functional tests into separate files - so we can test separately
