+++ tests/functional/test_info.py
@@ -1,0 +1,11 @@
+#
+# Test farm info method
+#
+
+def test_get_farm_info(test_farm):
+    info = test_farm.info()
+
+    assert 'name' in info
+    assert 'url' in info
+    assert 'api_version' in info
+    assert 'user' in info
